requirements.txt
# Core data science libraries
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Statistical analysis
statsmodels>=0.14.0
scikit-learn>=1.3.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
